Expand Unix tests to Android
Adjust the top-level imports to include Android, and add handling
for `testExitSignal()` and `testSubprocessPlatformOptionsCreateSession()`
on Android.

Move `ProcessControlTests.testSuspendResumeProcess()` into the Unix
test suite, since it was previously separated only due to the Unix
test suite not running on Android.
