+++ Tests/SubprocessTests/UnixTests.swift
@@ -9,13 +9,15 @@
 //
 //===----------------------------------------------------------------------===//
 
-#if canImport(Darwin) || canImport(Glibc)
+#if !os(Windows)
 import Foundation
 
-#if canImport(Glibc)
+#if canImport(Darwin)
+import Darwin
+#elseif canImport(Android)
+import Android
+#elseif canImport(Glibc)
 import Glibc
-#elseif canImport(Bionic)
-import Bionic
 #elseif canImport(Musl)
 import Musl
 #endif
@@ -156,6 +158,19 @@ extension SubprocessUnixTests {
         // platformOptions.createSession implies calls to setsid
         var platformOptions = PlatformOptions()
         platformOptions.createSession = true
+        #if os(Android)
+        // Android's `ps` doesn't support `-o pid,pgid,tpgid`. Read the shell's
+        // session fields directly from /proc instead. `$$` is the shell's own
+        // pid, which is the session and group leader after setsid; reading
+        // /proc/self/stat would observe `cat`, which is not the leader.
+        let statResult = try await Subprocess.run(
+            .path("/bin/sh"),
+            arguments: ["-c", "cat /proc/$$/stat"],
+            platformOptions: platformOptions,
+            output: .string(limit: .max)
+        )
+        try assertNewSessionCreated(fromProcStat: statResult)
+        #else
         // Check the process ID (pid), process group ID (pgid), and
         // controlling terminal's process group ID (tpgid)
         let psResult = try await Subprocess.run(
@@ -165,6 +180,7 @@ extension SubprocessUnixTests {
             output: .string(limit: .max)
         )
         try assertNewSessionCreated(with: psResult)
+        #endif
     }
 
     @Test(.requiresBash) func testTeardownSequence() async throws {
@@ -277,6 +293,60 @@ extension SubprocessUnixTests {
 
 // MARK: - Misc
 extension SubprocessUnixTests {
+    @Test(.timeLimit(.minutes(1)))
+    func testSuspendResumeProcess() async throws {
+        // Set up pipes manually so the test owns both ends. This lets us bound
+        // how long we wait for output. Standard sequence/iterator-based outputs
+        // use non-Sendable iterators that can't be moved into a child task.
+        let inputPipe = try FileDescriptor.pipe()
+        let outputPipe = try FileDescriptor.pipe()
+
+        // Make the read end non-blocking so readLine() can poll with a
+        // deadline instead of blocking indefinitely.
+        let flags = fcntl(outputPipe.readEnd.rawValue, F_GETFL)
+        try #require(fcntl(outputPipe.readEnd.rawValue, F_SETFL, flags | O_NONBLOCK) == 0)
+
+        try await outputPipe.readEnd.closeAfter {
+            try await inputPipe.writeEnd.closeAfter {
+                _ = try await Subprocess.run(
+                    .path("/bin/cat"),
+                    // cat reads from inputPipe.readEnd. The parent keeps writeEnd to feed it.
+                    input: .fileDescriptor(inputPipe.readEnd, closeAfterSpawningProcess: true),
+                    // cat writes to outputPipe.writeEnd. The parent keeps readEnd to drain it.
+                    output: .fileDescriptor(outputPipe.writeEnd, closeAfterSpawningProcess: true),
+                    error: .discarded
+                ) { subprocess in
+                    // Confirm cat is running and echoing before manipulating its state.
+                    try inputPipe.writeEnd.writeLine("ready")
+                    try #require(try await outputPipe.readEnd.readLine(timeout: .seconds(2)) == "ready")
+
+                    // Suspend cat, then write two lines it must not echo back. If
+                    // SIGSTOP took effect, no amount of waiting produces output,
+                    // and the lines accumulate in the pipe buffer until SIGCONT.
+                    try subprocess.send(signal: .suspend)
+                    try inputPipe.writeEnd.writeLine("first")
+                    try inputPipe.writeEnd.writeLine("second")
+
+                    // Give cat 200ms to (incorrectly) produce output. If the suspend
+                    // didn't work, cat already echoed and the pipe has bytes ready.
+                    let leaked = try await outputPipe.readEnd.readLine(timeout: .milliseconds(200))
+                    try #require(leaked == nil, "cat produced output while suspended: \(leaked ?? "")")
+
+                    // Resume cat. Both buffered lines must emerge in order, proving
+                    // the process accumulated state while stopped and released it
+                    // on SIGCONT.
+                    try subprocess.send(signal: .resume)
+                    try #require(try await outputPipe.readEnd.readLine(timeout: .seconds(2)) == "first")
+                    try #require(try await outputPipe.readEnd.readLine(timeout: .seconds(2)) == "second")
+
+                    // Tear down. SIGTERM makes cat exit; closeAfter closes the
+                    // parent's write end as cleanup once run() returns.
+                    try subprocess.send(signal: .terminate)
+                }
+            }
+        }
+    }
+
     @Test func testExitSignal() async throws {
         let signalsToTest: [CInt] = [SIGKILL, SIGTERM, SIGINT]
         for signal in signalsToTest {
@@ -285,7 +355,19 @@ extension SubprocessUnixTests {
                 arguments: ["-c", "kill -\(signal) $$"],
                 output: .discarded
             )
+            #if os(Android)
+            // When terminated by a catchable signal, Android's /bin/sh exits
+            // normally with a status of 128+n instead of dying by the signal.
+            // SIGKILL is uncatchable and still produces a signal-based termination.
+            // https://www.gnu.org/software/autoconf/manual/autoconf-2.69/html_node/Signal-Handling.html
+            let expected: TerminationStatus =
+                signal == SIGKILL
+                ? .signaled(signal)
+                : .exited(128 + signal)
+            #expect(result.terminationStatus == expected)
+            #else
             #expect(result.terminationStatus == .signaled(signal))
+            #endif
         }
     }
 
@@ -679,6 +761,84 @@ internal func assertNewSessionCreated(
     #expect(tpgid <= 0)
 }
 
+#if os(Android)
+internal func assertNewSessionCreated<Output: OutputProtocol>(
+    fromProcStat result: ExecutionResult<Void, StringOutput<UTF8>, Output>
+) throws {
+    #expect(result.terminationStatus.isSuccess)
+    let statLine = try #require(result.standardOutput)
+    // `comm` can contain spaces and parentheses, so bracket it by the first
+    // '(' and the last ')' rather than splitting the whole line on whitespace.
+    let openParen = try #require(
+        statLine.firstIndex(of: "("),
+        "/proc stat was in an unexpected format:\n\n\(statLine)"
+    )
+    let closeParen = try #require(
+        statLine.lastIndex(of: ")"),
+        "/proc stat was in an unexpected format:\n\n\(statLine)"
+    )
+    let pid = try #require(Int(statLine[..<openParen].trimmingCharacters(in: .whitespaces)))
+    // Fields after `comm`: [0] state, [1] ppid, [2] pgrp, [3] session, [4] tty_nr, [5] tpgid
+    let fields = statLine[statLine.index(after: closeParen)...].split(separator: " ")
+    try #require(fields.count >= 6, "/proc stat was in an unexpected format:\n\n\(statLine)")
+    let pgid = try #require(Int(fields[2]))
+    let session = try #require(Int(fields[3]))
+    let tpgid = try #require(Int(fields[5]))
+    #expect(pid == pgid)
+    #expect(pid == session)
+    #expect(tpgid <= 0)
+}
+#endif
+
+extension FileDescriptor {
+    /// Writes a line plus newline to the file descriptor.
+    fileprivate func writeLine(_ line: String) throws {
+        let bytes = Array("\(line)\n".utf8)
+        try bytes.withUnsafeBufferPointer { ptr in
+            _ = try self.write(UnsafeRawBufferPointer(ptr))
+        }
+    }
+
+    /// Reads a single line (up to and excluding the next `\n`) from a
+    /// non-blocking file descriptor, returning `nil` if no line arrives
+    /// within `timeout`.
+    fileprivate func readLine(timeout: Duration) async throws -> String? {
+        let deadline = ContinuousClock.now + timeout
+        var accumulated: [UInt8] = []
+
+        while ContinuousClock.now < deadline {
+            var byte: UInt8 = 0
+            let n = withUnsafeMutablePointer(to: &byte) { ptr in
+                #if canImport(Darwin)
+                return Darwin.read(self.rawValue, ptr, 1)
+                #elseif canImport(Android)
+                return Android.read(self.rawValue, ptr, 1)
+                #elseif canImport(Glibc)
+                return Glibc.read(self.rawValue, ptr, 1)
+                #elseif canImport(Musl)
+                return Musl.read(self.rawValue, ptr, 1)
+                #endif
+            }
+
+            if n == 1 {
+                if byte == 0x0A {
+                    return String(decoding: accumulated, as: UTF8.self)
+                }
+                accumulated.append(byte)
+            } else if n == 0 {
+                return accumulated.isEmpty ? nil : String(decoding: accumulated, as: UTF8.self)
+            } else if errno == EAGAIN || errno == EWOULDBLOCK {
+                try await Task.sleep(for: .milliseconds(5))
+            } else if errno == EINTR {
+                continue
+            } else {
+                throw Errno(rawValue: errno)
+            }
+        }
+        return nil
+    }
+}
+
 // MARK: - Performance Tests
 extension SubprocessUnixTests {
     #if SubprocessFoundation
@@ -738,4 +898,4 @@ extension SubprocessUnixTests {
     #endif
 }
 
-#endif // canImport(Darwin) || canImport(Glibc)
+#endif // !os(Windows)
